requirements.txt
starlette==0.38.2
uvicorn==0.30.5
python-binance
binance-connector==3.8.1 # https://binance-connector.readthedocs.io/en/latest/getting_started.html
ipython==8.26.0
jinja2==3.1.4
python-multipart==0.0.9
pandas